resize code refactoring
