+++ src/components/table/table.functions.js
@@ -1,0 +1,3 @@
+export function isResize(event) {
+  return event.target.dataset.resize ? true : false
+}
